+++ src/ApexRacers.Data/Migrations/20260814003047_AddDriverCustIdToPersonalLap.cs
@@ -1,0 +1,30 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ApexRacers.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddDriverCustIdToPersonalLap : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<long>(
+                name: "DriverCustId",
+                schema: "iracing",
+                table: "PersonalLaps",
+                type: "bigint",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "DriverCustId",
+                schema: "iracing",
+                table: "PersonalLaps");
+        }
+    }
+}